Improvements to documentation and handling values spec'd in asset file
