+++ modules/space/rendering/renderablesmallbody.h
@@ -38,6 +38,8 @@
 
 namespace openspace {
 
+namespace documentation { struct Documentation; }
+
 class RenderableSmallBody : public RenderableOrbitalKepler {
 public:
     RenderableSmallBody(const ghoul::Dictionary& dictionary);
@@ -52,13 +54,16 @@ private:
 
     std::vector<std::string> _sbNames;
     std::function<void()> _updateContiguousModeSelect;
+    std::function<void()> _updateRenderUpperLimitSelect;
 
     /// The index array that is potentially used in the draw call. If this is empty, no
     /// element draw call is used.
     std::vector<unsigned int> _indexBufferData;
     bool contiguousMode = true;
     properties::BoolProperty _contiguousMode;
+    properties::UIntProperty _upperLimit;
     properties::Property::OnChangeHandle _contiguousModeCallbackhandle;
+    properties::Property::OnChangeHandle _upperLimitCallbackHandle;
 };
 
 } // namespace openspace
